Isolate historical private-storage paths behind storage boundaries
